@@ -1,16 +1,32 @@
 apply plugin: 'com.android.application'
 apply plugin: 'com.google.gms.google-services'
+apply plugin: 'com.google.firebase.crashlytics'
 
 android {
-    compileSdkVersion 28
+    signingConfigs {
+        release {
+            storeFile file('C:\\Users\\ibindra\\AndroidStudioProjects\\JustBakers\\gradle\\keystore.jks')
+            storePassword 'justbakers'
+            keyPassword 'justbakers'
+            keyAlias = 'key0'
+        }
+
+        debug {
+            storeFile file('C:\\Users\\ibindra\\AndroidStudioProjects\\JustBakers\\gradle\\keystore.jks')
+            storePassword 'justbakers'
+            keyPassword 'justbakers'
+            keyAlias = 'key0'
+        }
+    }
+    compileSdkVersion 29
 
 
     defaultConfig {
         applicationId "com.ecom.justbakers"
         minSdkVersion 17
-        targetSdkVersion 28
-        versionCode 11
-        versionName "1.1"
+        targetSdkVersion 29
+        versionCode 16
+        versionName "1.6"
         multiDexEnabled true
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
