@@ -1,11 +1,10 @@
 package com.ecom.justbakers;
 
+import android.app.Activity;
 import android.app.ProgressDialog;
 import android.content.Context;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.res.Resources;
-import android.database.DataSetObserver;
 import android.graphics.PorterDuff;
 import android.os.Bundle;
 import android.text.InputType;
@@ -13,12 +12,9 @@ import android.util.DisplayMetrics;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
-import android.view.ViewGroup;
-import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.LinearLayout;
-import android.widget.ListAdapter;
 import android.widget.ListView;
 import android.widget.ProgressBar;
 import android.widget.TableRow;
@@ -27,11 +23,12 @@ import android.widget.Toast;
 
 import com.ecom.justbakers.Adapters.CustomOrderListAdapter;
 import com.ecom.justbakers.Adapters.CustomProductListAdapter;
+import com.ecom.justbakers.Adapters.CustomerPendingOrderListAdapter;
 import com.ecom.justbakers.Adapters.OrderDetailsListAdapter;
 import com.ecom.justbakers.Classes.BargainProductClass;
 import com.ecom.justbakers.Classes.ProductClass;
-import com.ecom.justbakers.MenuItemListeners.MenuBargainItemListener;
 import com.ecom.justbakers.MenuItemListeners.MenuCartItemListener;
+import com.ecom.justbakers.orders.InfoClass;
 import com.ecom.justbakers.orders.OrderClass;
 import com.firebase.client.DataSnapshot;
 import com.firebase.client.Firebase;
@@ -44,6 +41,7 @@ import com.google.android.material.navigation.NavigationView;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
+import java.util.Date;
 import java.util.Random;
 
 import androidx.appcompat.app.ActionBarDrawerToggle;
@@ -66,44 +64,63 @@ public class UserActivity extends AppCompatActivity implements NavigationView.On
     private String userId = null;
     CustomProductListAdapter productAdapter;
     CustomOrderListAdapter orderAdapter;
-    static  TextView cartcounterTV;
+    CustomerPendingOrderListAdapter customerPendingOrderAdapter;
+    TextView cartcounterTV;
     TextView bargaincounterTV;
-    private boolean pendingOrderFlag = true;
+    private boolean adminLogin = false;
+    /*private boolean pendingOrderFlag = true;
     private  boolean placedOrderFlag = false;
-    private  boolean deliveredOrderFlag = false;
+    private  boolean deliveredOrderFlag = false;*/
 
     ValueEventListener cartValueEventListener;
     ValueEventListener prodValueEventListener;
-    /** MAKING A REFERENCE TO PRODUCT DISPLAY FIREBASE **/
+    ValueEventListener placedValueEventListener;
+    ValueEventListener custValueEventListener;
+
+    /* MAKING A REFERENCE TO PRODUCT DISPLAY FIREBASE */
     Firebase productRef = new Firebase("https://justbakers-285be.firebaseio.com/productDisplay");
     /** MAKING A REFERENCE TO CART URL IN FIREBASE FOR THE VALUES TO BE PUSHED **/
     Firebase cartRef = new Firebase("https://justbakers-285be.firebaseio.com/")  // https://justbakers-285be.firebaseio.com/
             .child("customers");
-    /** MAKING A REFERENCE TO BARGAIN REQUESTS CART **/
+    /* MAKING A REFERENCE TO BARGAIN REQUESTS CART */
     //  Firebase bargaincartRef = new Firebase("https://justbakers-285be.firebaseio.com/")
     //         .child("BargainCarts");
     /** MAKING A REFERENCE TO ORDERS LIST FIREBASE **/
     Firebase placedOrderRef = new Firebase("https://justbakers-285be.firebaseio.com/customers");
     Firebase deliveredOrderRef = new Firebase("https://justbakers-285be.firebaseio.com/customers");
+    Firebase custRef =null;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_user);
         Firebase.setAndroidContext(this);
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+        Toolbar toolbar = findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
-        productList =(ListView)findViewById(R.id.ProductList);
 
 
-        /**CHECK IF IN ANY CASE USERNAME STRING IS NULL OR EMPTY GO BACK TO LOGIN ACTIVITY**/
-        if(LoginActivity.getDefaults("UserID",this)==null || LoginActivity.getDefaults("UserID",this).equals(""))
+        productList = findViewById(R.id.ProductList);
+
+        userId = LoginActivity.getDefaults("UserID",this);
+        String gmail = LoginActivity.getDefaults("Gmail",this);
+        DrawerLayout drawer = findViewById(R.id.drawer_layout);
+        NavigationView navigationView = findViewById(R.id.nav_view);
+        /* CHECK IF IN ANY CASE USERNAME STRING IS NULL OR EMPTY GO BACK TO LOGIN ACTIVITY */
+        if(userId ==null || LoginActivity.getDefaults("UserID",this).equals(""))
         {
             Intent intent = new Intent(this,LoginActivity.class);
             this.startActivity(intent);
         }
         else {
-            userId = LoginActivity.getDefaults("UserID",this);
+
+            if (gmail.equals(LoginActivity.getAdminUser())) {
+                adminLogin = true;
+                custRef = new Firebase("https://justbakers-285be.firebaseio.com/pendingOrders");
+
+                MenuItem customerMenuItem = navigationView.getMenu().findItem(R.id.nav_adminview);
+                customerMenuItem.setVisible(true);
+              //  customerView.setVisibility(View.VISIBLE);
+            }
 
             cartRef = new Firebase("https://justbakers-285be.firebaseio.com/customers/"+userId+"/orders/pending")  // https://justbakers-285be.firebaseio.com/
                     .child("cart");
@@ -114,23 +131,26 @@ public class UserActivity extends AppCompatActivity implements NavigationView.On
 
         }
 
-        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
+
         ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(
                 this, drawer, toolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close);
         drawer.setDrawerListener(toggle);
         toggle.syncState();
 
-        /**GETTING THE NAVIGATION BAR'S HEADER VIEW**/
-        NavigationView navigationView = (NavigationView) findViewById(R.id.nav_view);
+        /* GETTING THE NAVIGATION BAR'S HEADER VIEW */
+
         View header = navigationView.getHeaderView(0);
+        NavigationView customerView = navigationView.findViewById(R.id.nav_adminview);
 
-        /** SETTING THE TEXTS IN NAVIGATION HEADER **/
-        TextView Email = (TextView) header.findViewById(R.id.Email);
-        Email.setText(LoginActivity.getDefaults("UserID",this));
-        TextView Name = (TextView) header.findViewById(R.id.Name);
-        Name.setText(LoginActivity.getDefaults("UserID",this));
 
-        /**RANDOM TIPS GENERATION**/
+
+        /* SETTING THE TEXTS IN NAVIGATION HEADER */
+        TextView Email = header.findViewById(R.id.Email);
+        Email.setText(LoginActivity.getDefaults("Gmail",this)+"@gmail.com");
+        TextView Name = header.findViewById(R.id.Name);
+        Name.setText(LoginActivity.getDefaults("Gmail",this));
+
+        /* RANDOM TIPS GENERATION**/
 //        TextView UserTips = (TextView) findViewById(R.id.UserTips);
         Resources res = getResources();
 
@@ -140,14 +160,15 @@ public class UserActivity extends AppCompatActivity implements NavigationView.On
         String tip = userTips[random.nextInt(userTips.length)];
         //UserTips.setText(tip);
 
-        /** CODE FOR SIGN OUT IN NAVIGATION HEADER **/
-        Button SignOut = (Button) header.findViewById(R.id.signout);
-        SignOut.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent signoutintent = new Intent(UserActivity.this, LoginActivity.class);
-                UserActivity.this.startActivity(signoutintent);
-            }
+        /* CODE FOR SIGN OUT IN NAVIGATION HEADER */
+        Button SignOut = header.findViewById(R.id.signout);
+        SignOut.setOnClickListener(v -> {
+            LoginActivity.getGoogleSignInClient().signOut()
+                    .addOnCompleteListener((Activity) getApplicationContext(), task -> {
+                        // ...
+                    });
+            Intent signoutintent = new Intent(UserActivity.this, LoginActivity.class);
+            UserActivity.this.startActivity(signoutintent);
         });
         navigationView.setNavigationItemSelectedListener(this);
     }
@@ -155,20 +176,73 @@ public class UserActivity extends AppCompatActivity implements NavigationView.On
     @Override
     protected void onStart() {
         super.onStart();
-        /** INITIALISATION **/
+
+        /* INITIALISATION */
         ArrayList<ProductClass> productClassList = new ArrayList<>();
+        ArrayList<InfoClass> customersInfoList = new ArrayList<>();
         ArrayList<ProductClass> cartProductClassList = new ArrayList<>();
         ArrayList<OrderClass> placedOrdersClassList = new ArrayList<>();
+        ArrayList<OrderClass> customersOrderList = new ArrayList<>();
 
         //      FirebaseBargainCartCount(bargaincartRef);
 
-        /** CREATING AND STARTING THE PROGRESS BAR **/
-        progress=new ProgressDialog(this);
+        /* CREATING AND STARTING THE PROGRESS BAR */
+        progress = new ProgressDialog(this);
         progress.setMessage(getResources().getString(R.string.Loading1));
         progress.setProgressStyle(ProgressDialog.STYLE_SPINNER);
         progress.setIndeterminate(true);
         progress.setCancelable(false);
         progress.show();
+
+
+        if (adminLogin) {
+
+            custValueEventListener = new ValueEventListener() {
+                @Override
+                public void onDataChange(DataSnapshot dataSnapshot) {
+                    for (DataSnapshot postSnapshot : dataSnapshot.getChildren()) {
+                        InfoClass info = (postSnapshot.child ("info")).getValue(InfoClass.class);
+                        DataSnapshot ds1 = postSnapshot.child ("order");
+                        String ordernumber = (String)(ds1.child("ordernumber")).getValue();
+                        double finalamount = (double)(ds1.child("receipt").child("finalamount")).getValue();
+                        String status = (String)(ds1.child("status")).getValue();
+                        long time = (long)(ds1.child("time")).getValue();
+                        long dt = (long)(ds1.child("date")).getValue();
+                        Date date = new Date(dt);
+                        OrderClass.OrderReceiptClass receipt = new OrderClass.OrderReceiptClass(finalamount);
+                        ArrayList<ProductClass> cart = new ArrayList<>();
+                        for ( DataSnapshot childSnapShot :  ds1.child("cart").getChildren()) {
+
+                          ProductClass pc = childSnapShot.getValue(ProductClass.class);
+                          cart.add(pc);
+                        }
+
+
+                        customersInfoList.add(info);
+
+                        OrderClass order = new OrderClass(ordernumber, cart, receipt, date, time, status);
+                        customersOrderList.add(order);
+
+                    }
+                }
+
+                @Override
+                public void onCancelled(FirebaseError firebaseError) {
+
+                }
+            };
+
+            custRef.addValueEventListener(custValueEventListener);
+            customerPendingOrderAdapter = new CustomerPendingOrderListAdapter(UserActivity.this, customersInfoList, customersOrderList);
+            productList.setAdapter(customerPendingOrderAdapter);
+
+        }
+
+
+
+
+
+
 
         cartValueEventListener = new ValueEventListener() {
             @Override
@@ -179,7 +253,7 @@ public class UserActivity extends AppCompatActivity implements NavigationView.On
                         System.out.println(post.getName() + " - " + post.getDescription() + "-" + post.getId());
                         ProductClass currentProduct = new ProductClass(post.getName()
                                 , post.getImage()
-                                , post.getprice()
+                                , post.getPrice()
                                 , post.getDescription()
                                 , post.getSeller()
                                 , post.getId()
@@ -201,7 +275,7 @@ public class UserActivity extends AppCompatActivity implements NavigationView.On
 
         cartRef.addValueEventListener(cartValueEventListener);
 
-        /** ADDING THE VALUE EVENT LISTENER TO PRODUCTREF FOR DISPLAYING THE LIST OF PRODUCTS **/
+        /* ADDING THE VALUE EVENT LISTENER TO PRODUCTREF FOR DISPLAYING THE LIST OF PRODUCTS */
 
         productAdapter = new CustomProductListAdapter(UserActivity.this
                 , productClassList, cartProductClassList
@@ -210,17 +284,17 @@ public class UserActivity extends AppCompatActivity implements NavigationView.On
                 (position, v) -> {
                     //Custom Toast Display Function defined below - Displays custom toast message
                     //     CustomToastDisplay(UserActivity.this,getResources().getString(R.string.SuccessfulCart));
-                    /**FUNCTION DEFINED IN THE LAST- PUSHING THE PRODUCT TO CART DATABASE**/
+                    /* FUNCTION DEFINED IN THE LAST- PUSHING THE PRODUCT TO CART DATABASE */
                     cartButtonClick(position, v, productClassList, cartRef);
                 } // BARGAIN BUTTON ON CLICK LISTENER
                 , (position, v) -> {
-            /**FUNCTION DEFINED AT THE END OF THE CLASS, SHOW AN DIALOG ASKING FOR BARGAIN PRICE AND PUSH IT TO FIRBASE*/
+            /* FUNCTION DEFINED AT THE END OF THE CLASS, SHOW AN DIALOG ASKING FOR BARGAIN PRICE AND PUSH IT TO FIRBASE */
             //  BargainButtonAlertDialog(UserActivity.this,position,v,ProductClassList,bargaincartRef);
         }, //ADD ITEM BUTTON ON CLICK LISTENER
                 (position, v) -> {
                     //Custom Toast Display Function defined below - Displays custom toast message
                     //    CustomToastDisplay(UserActivity.this,getResources().getString(R.string.SuccessfulCart));
-                    /**FUNCTION DEFINED IN THE LAST- PUSHING THE PRODUCT TO CART DATABASE**/
+                    /* FUNCTION DEFINED IN THE LAST- PUSHING THE PRODUCT TO CART DATABASE */
                     addItemButtonClick(position, v, productClassList, cartRef);
                 }
                 , //* SUB ITEM BUTTON ON CLICK LISTENER
@@ -242,7 +316,7 @@ public class UserActivity extends AppCompatActivity implements NavigationView.On
                         System.out.println(post.getName() + " - " + post.getDescription() + "-" + post.getId());
                         ProductClass currentProduct = new ProductClass(post.getName()
                                 , post.getImage()
-                                , post.getprice()
+                                , post.getPrice()
                                 , post.getDescription()
                                 , post.getSeller()
                                 , post.getId()
@@ -266,35 +340,26 @@ public class UserActivity extends AppCompatActivity implements NavigationView.On
 
         productRef.addValueEventListener(prodValueEventListener);
 
-        /** ADDING THE VALUE EVENT LISTENER TO PRODUCTREF FOR DISPLAYING THE LIST OF PRODUCTS **/
+        /* ADDING THE VALUE EVENT LISTENER TO PRODUCTREF FOR DISPLAYING THE LIST OF PRODUCTS **/
         orderAdapter = new CustomOrderListAdapter(UserActivity.this
                 , placedOrdersClassList
-                ,/** ORDER DETAILS BUTTON ON CLICK LISTENER **/
-                new CustomOrderListAdapter.OrderButtonClickListener(){
-                    @Override
-                    public void onButtonClick(int position,View v) {
-                        /**FUNCTION DEFINED IN THE LAST- PUSHING THE PRODUCT TO CART DATABASE**/
-                        orderTrackButtonClick(position,v, placedOrdersClassList, placedOrderRef);
-                    }
+                ,/* ORDER DETAILS BUTTON ON CLICK LISTENER */
+                (position, v) -> {
+                    /*FUNCTION DEFINED IN THE LAST- PUSHING THE PRODUCT TO CART DATABASE*/
+                    orderTrackButtonClick(position,v, placedOrdersClassList, placedOrderRef);
                 }
-                ,/** ORDER DETAILS BUTTON ON CLICK LISTENER **/
-                new CustomOrderListAdapter.OrderButtonClickListener(){
-                    @Override
-                    public void onButtonClick(int position,View v) {
-                        /**FUNCTION DEFINED IN THE LAST- PUSHING THE PRODUCT TO CART DATABASE**/
-                        orderDetailsButtonClick(position,v, placedOrdersClassList, placedOrderRef);
-                    }
+                ,/* ORDER DETAILS BUTTON ON CLICK LISTENER */
+                (position, v) -> {
+                    /*FUNCTION DEFINED IN THE LAST- PUSHING THE PRODUCT TO CART DATABASE*/
+                    orderDetailsButtonClick(position,v, placedOrdersClassList, placedOrderRef);
                 }
-                ,/** CANCEL ORDER BUTTON ON CLICK LISTENER **/
-                new CustomOrderListAdapter.OrderButtonClickListener(){
-                    @Override
-                    public void onButtonClick(int position,View v) {
-                        /**FUNCTION DEFINED IN THE LAST- PUSHING THE PRODUCT TO CART DATABASE**/
-                        cancelOrderButtonClick(position,v, placedOrdersClassList);
-                    }
+                ,/* CANCEL ORDER BUTTON ON CLICK LISTENER **/
+                (position, v) -> {
+                    /*FUNCTION DEFINED IN THE LAST- PUSHING THE PRODUCT TO CART DATABASE*/
+                    cancelOrderButtonClick(position,v, placedOrdersClassList);
                 });
 
-        placedOrderRef.addValueEventListener(new ValueEventListener() {
+            placedValueEventListener = new ValueEventListener() {
             @Override
             public void onDataChange(DataSnapshot dataSnapshot) {
                 if( !UserActivity.this.isFinishing() && UserActivity.this.getLifecycle().getCurrentState().isAtLeast(Lifecycle.State.STARTED) ) {
@@ -321,8 +386,10 @@ public class UserActivity extends AppCompatActivity implements NavigationView.On
             public void onCancelled(FirebaseError firebaseError) {
                 System.out.println("The read failed: " + firebaseError.getMessage());
             }
-        });
-/**
+        };
+
+        placedOrderRef.addValueEventListener( placedValueEventListener);
+/*
  DeliveredOrderRef.addValueEventListener(new ValueEventListener() {
 @Override
 public void onDataChange(DataSnapshot dataSnapshot) {
@@ -361,7 +428,7 @@ System.out.println("The read failed: " + firebaseError.getMessage());
 }
 });
 
- **/
+ */
     }
 
     @Override
@@ -432,11 +499,11 @@ System.out.println("The read failed: " + firebaseError.getMessage());
 
     @Override
     public void onBackPressed() {
-        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
+        DrawerLayout drawer = findViewById(R.id.drawer_layout);
         if (drawer.isDrawerOpen(GravityCompat.START)) {
             drawer.closeDrawer(GravityCompat.START);
         } else {
-            super.onBackPressed();
+            //super.onBackPressed();
         }
     }
 
@@ -455,6 +522,7 @@ System.out.println("The read failed: " + firebaseError.getMessage());
         super.onStop();
         cartRef.removeEventListener(cartValueEventListener);
         productRef.removeEventListener(prodValueEventListener);
+        placedOrderRef.removeEventListener(placedValueEventListener);
 
     }
 
@@ -471,10 +539,10 @@ System.out.println("The read failed: " + firebaseError.getMessage());
         getMenuInflater().inflate(R.menu.user, menu);
 
         View menu_item_cart = menu.findItem(R.id.action_cart).getActionView();
-        cartcounterTV = (TextView) menu_item_cart.findViewById(R.id.cartcounter);
+        cartcounterTV = menu_item_cart.findViewById(R.id.cartcounter);
        // bargaincounterTV = (TextView) menu_item_jb.findViewById(R.id.jbcounter);
 
-        /**UPDATING INITIAL THE CART COUNTS AND ONMENUITEMLISTENER FOR MENU_WITTY ITEM**/
+        /*UPDATING INITIAL THE CART COUNTS AND ONMENUITEMLISTENER FOR MENU_WITTY ITEM**/
 
         updateCartCount(productAdapter.getCartItemsCount());
 
@@ -486,7 +554,7 @@ System.out.println("The read failed: " + firebaseError.getMessage());
             }
         };*/
 
-        /**UPDATING INITIAL THE BARGAIN REQUEST COUNTS AND ONMENUITEMLISTENER FOR MENU_CART ITEM**/
+        /*UPDATING INITIAL THE BARGAIN REQUEST COUNTS AND ONMENUITEMLISTENER FOR MENU_CART ITEM**/
         updateBargainCartCount(bargainCartIndex);
         MenuCartItemListener cartItemListener = new MenuCartItemListener(menu_item_cart, "Show message") {
             @Override
@@ -522,42 +590,35 @@ System.out.println("The read failed: " + firebaseError.getMessage());
         if (id == R.id.nav_placedorders) {
             llContact.setVisibility(View.GONE);
             productList.setVisibility(View.VISIBLE);
-
-
             orderTableHeader.setVisibility(View.VISIBLE);
             setActionBarTitle("Your Orders");
-
             productList.setAdapter(orderAdapter);
-
-            setFlags(false, true, false);
         } else if (id == R.id.nav_yourcart) {
             llContact.setVisibility(View.GONE);
             productList.setVisibility(View.VISIBLE);
-
-
             orderTableHeader.setVisibility(View.GONE);
             setActionBarTitle("Your Cart");
-
-
-            /**ONCLICKLISTENER ON THE CARDS-PASSING THE DETAILS VIA INTENT TO DESCRIPTION ACTIVITY**/
+            /*ONCLICKLISTENER ON THE CARDS-PASSING THE DETAILS VIA INTENT TO DESCRIPTION ACTIVITY**/
 
             productList.setOnItemClickListener((parent, view, position, id1) -> {
-                ProductClass CurrentProduct = (ProductClass) productAdapter.getItem(position);
+                ProductClass CurrentProduct = productAdapter.getItem(position);
                 Intent DescriptionIntent = new Intent(UserActivity.this, DescriptionActivity.class);
                 DescriptionIntent.putExtra("ProductDetails",CurrentProduct);
                 UserActivity.this.startActivity(DescriptionIntent);
             });
-
             productList.setAdapter(productAdapter);
-            setFlags(true, false, false);
-        }  else if (id == R.id.nav_contact) {
-
+        }  if (id == R.id.nav_adminview) {
+            llContact.setVisibility(View.GONE);
+            productList.setVisibility(View.VISIBLE);
+            setActionBarTitle("Customer Orders");
+            productList.setAdapter(customerPendingOrderAdapter);
+        }else if (id == R.id.nav_contact) {
             llContact.setVisibility(View.VISIBLE);
             productList.setVisibility(View.GONE);
             orderTableHeader.setVisibility(View.GONE);
-
         }
-        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
+
+        DrawerLayout drawer = findViewById(R.id.drawer_layout);
         drawer.closeDrawer(GravityCompat.START);
         return true;
     }
@@ -596,15 +657,12 @@ System.out.println("The read failed: " + firebaseError.getMessage());
     public void updateCartCount(final int new_number) {
         //cartIndex = new_number;
         if (cartcounterTV == null) return;
-        runOnUiThread(new Runnable() {
-            @Override
-            public void run() {
-                if (new_number == 0)
-                    cartcounterTV.setVisibility(View.INVISIBLE);
-                else {
-                    cartcounterTV.setVisibility(View.VISIBLE);
-                    cartcounterTV.setText(Integer.toString(new_number));
-                }
+        runOnUiThread(() -> {
+            if (new_number == 0)
+                cartcounterTV.setVisibility(View.INVISIBLE);
+            else {
+                cartcounterTV.setVisibility(View.VISIBLE);
+                cartcounterTV.setText(Integer.toString(new_number));
             }
         });
     }
@@ -613,33 +671,30 @@ System.out.println("The read failed: " + firebaseError.getMessage());
     public void updateBargainCartCount(final int new_number) {
         bargainCartIndex = new_number;
         if (bargaincounterTV == null) return;
-        runOnUiThread(new Runnable() {
-            @Override
-            public void run() {
-                if (new_number == 0)
-                    bargaincounterTV.setVisibility(View.INVISIBLE);
-                else {
-                    bargaincounterTV.setVisibility(View.VISIBLE);
-                    bargaincounterTV.setText(Integer.toString(new_number));
-                }
+        runOnUiThread(() -> {
+            if (new_number == 0)
+                bargaincounterTV.setVisibility(View.INVISIBLE);
+            else {
+                bargaincounterTV.setVisibility(View.VISIBLE);
+                bargaincounterTV.setText(Integer.toString(new_number));
             }
         });
     }
 
     /**FUNCTION TO SHOW CUSTOM TOAST MESSAGE**/
     public static void customToastDisplay(Context context, String Message){
-        /** CUSTOM TOAST MESSAGE **/
+        /* CUSTOM TOAST MESSAGE **/
         Toast toast = Toast.makeText(context,"", Toast.LENGTH_SHORT);
         View view = toast.getView();
         view.setBackgroundResource(R.drawable.rounded_square);
-        TextView text = (TextView) view.findViewById(android.R.id.message);
+        TextView text = view.findViewById(android.R.id.message);
         text.setTextColor(context.getResources().getColor(R.color.colorWhite));
         text.setText(Message);
         toast.show();
     }
 
     public void cartButtonClick(int position, View v, ArrayList<ProductClass> productList , Firebase ref){
-        TextView addedInfo = (TextView) v.findViewById(R.id.addedInfo);
+        TextView addedInfo = v.findViewById(R.id.addedInfo);
         addedInfo.setVisibility(View.VISIBLE);
 
         //updateCartCount(++cartIndex);
@@ -648,26 +703,27 @@ System.out.println("The read failed: " + firebaseError.getMessage());
         product.setQuantity(1);
         ProductClass CartProduct = new ProductClass(product.getName()
                 ,product.getImage()
-                ,product.getprice()
+                ,product.getPrice()
                 ,product.getDescription()
                 ,product.getSeller()
                 ,product.getId()
                 ,product.getQuantity()
                 ,product.getLimit());
-        /** PUSHING THE PRODUCT TO THE CART DATABASE **/
+        /* PUSHING THE PRODUCT TO THE CART DATABASE **/
         ref.child(CartProduct.getId()).setValue(CartProduct);
         addedInfo.setText("This product is in your cart. Max quantity : " + CartProduct.getLimit());
         //THIS FUNCTION WILL UPDATE THE CART COUNTER NODE IN OUR DATABASE
         updateCartCounterInFirebase();
-        info.hoang8f.widget.FButton addCart = (info.hoang8f.widget.FButton) v.findViewById(R.id.Addtocart);
-        info.hoang8f.widget.FButton addItem = (info.hoang8f.widget.FButton) v.findViewById(R.id.addbutton);
-        info.hoang8f.widget.FButton subItem = (info.hoang8f.widget.FButton) v.findViewById(R.id.subbutton);
-        info.hoang8f.widget.FButton quantity = (info.hoang8f.widget.FButton) v.findViewById(R.id.quantitybutton);
+        info.hoang8f.widget.FButton addCart = v.findViewById(R.id.Addtocart);
+        info.hoang8f.widget.FButton addItem = v.findViewById(R.id.addbutton);
+        info.hoang8f.widget.FButton subItem = v.findViewById(R.id.subbutton);
+        info.hoang8f.widget.FButton quantity = v.findViewById(R.id.quantitybutton);
         addItem.setVisibility(View.VISIBLE);
         subItem.setVisibility(View.VISIBLE);
         quantity.setVisibility(View.VISIBLE);
         addCart.setVisibility(View.GONE);
         quantity.setText("1");
+
 
 
     }
@@ -747,7 +803,7 @@ System.out.println("The read failed: " + firebaseError.getMessage());
         cancel_order_button.setVisibility(vis);
 
         for ( ProductClass pc : curOrder.getCart()) {
-            cart_items += "Name : " + pc.getName()  + " Quantity : " + pc.getQuantity() + "  Price : " + pc.getprice() + "\n";
+            cart_items += "Name : " + pc.getName()  + " Quantity : " + pc.getQuantity() + "  Price : " + pc.getPrice() + "\n";
         }
 
         String fulldate =  curOrder.getDate().toString();
@@ -800,7 +856,7 @@ System.out.println("The read failed: " + firebaseError.getMessage());
         return curOrder;
     }
     public void bargainButtonAlertDialog(Context context, final int position, View v, final ArrayList<ProductClass> ProductList , final Firebase ref){
-        final TextView addedInfo = (TextView) v.findViewById(R.id.addedInfo);
+        final TextView addedInfo = v.findViewById(R.id.addedInfo);
         AlertDialog.Builder builder = new AlertDialog.Builder(UserActivity.this);
         builder.setTitle(R.string.BargainQuestion);
         // SET UP THE INPUT
@@ -812,58 +868,51 @@ System.out.println("The read failed: " + firebaseError.getMessage());
         builder.setView(input);
 
         // SETUP THE BUTTONS
-        builder.setPositiveButton("CONFIRM", new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
-                ProductClass bargainproduct = ProductList.get(position);
-                // A CHECK IF THE BARGAIN PRICE REQUESTED IS NOT LESS THAN 40% OR TWICE OF PRODUCT PRICE
-                if((Integer.parseInt(input.getText().toString()))< (0.4*bargainproduct.getprice()) || (Integer.parseInt(input.getText().toString()))> (2*bargainproduct.getprice())){
-                    customToastDisplay(UserActivity.this,getResources().getString(R.string.bargainError));
-                }
-                else{
-                    BargainProductClass BargainProduct = new BargainProductClass(bargainproduct.getName()
-                            ,bargainproduct.getImage()
-                            ,bargainproduct.getprice()
-                            ,Integer.parseInt(input.getText().toString())
-                            ,bargainproduct.getDescription()
-                            ,bargainproduct.getSeller()
-                            ,0
-                            ,bargainproduct.getId()
-                            ,""
-                            ,0
-                            ,1
-                            ,System.currentTimeMillis());
-                    /** PUSHING THE BARGAINPRODUCT TO THE BARGAINREQUESTS DATABASE **/
-                    ref.push().setValue(BargainProduct);
-                    //THIS FUNCTION WILL UPDATE THE BARGAIN REQUEST COUNTER NODE IN OUR DATABASE
-                    updateBargainCartCounterInFirebase();
-                    bargainCartIndex = bargainCartIndex + 1;
-                    updateBargainCartCount(bargainCartIndex);
-                    //MAKING ADDED INFO TEXT VIEW VISIBLE
-                    addedInfo.setVisibility(View.VISIBLE);
-                    addedInfo.setText("Bargain Requested. Click 'W' icon on the top for more details");
-                }
+        builder.setPositiveButton("CONFIRM", (dialog, which) -> {
+            ProductClass bargainproduct = ProductList.get(position);
+            // A CHECK IF THE BARGAIN PRICE REQUESTED IS NOT LESS THAN 40% OR TWICE OF PRODUCT PRICE
+            if((Integer.parseInt(input.getText().toString()))< (0.4*bargainproduct.getPrice()) || (Integer.parseInt(input.getText().toString()))> (2*bargainproduct.getPrice())){
+                customToastDisplay(UserActivity.this,getResources().getString(R.string.bargainError));
+            }
+            else{
+                BargainProductClass BargainProduct = new BargainProductClass(bargainproduct.getName()
+                        ,bargainproduct.getImage()
+                        ,bargainproduct.getPrice()
+                        ,Integer.parseInt(input.getText().toString())
+                        ,bargainproduct.getDescription()
+                        ,bargainproduct.getSeller()
+                        ,0
+                        ,bargainproduct.getId()
+                        ,""
+                        ,0
+                        ,1
+                        ,System.currentTimeMillis());
+                /* PUSHING THE BARGAINPRODUCT TO THE BARGAINREQUESTS DATABASE **/
+                ref.push().setValue(BargainProduct);
+                //THIS FUNCTION WILL UPDATE THE BARGAIN REQUEST COUNTER NODE IN OUR DATABASE
+                updateBargainCartCounterInFirebase();
+                bargainCartIndex = bargainCartIndex + 1;
+                updateBargainCartCount(bargainCartIndex);
+                //MAKING ADDED INFO TEXT VIEW VISIBLE
+                addedInfo.setVisibility(View.VISIBLE);
+                addedInfo.setText("Bargain Requested. Click 'W' icon on the top for more details");
             }
         });
-        builder.setNegativeButton("CANCEL", new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
-                dialog.cancel();
-            }
-        });
+        builder.setNegativeButton("CANCEL", (dialog, which) -> dialog.cancel());
         builder.show();
     }
 
     public void addItemButtonClick(int position, View v, ArrayList<ProductClass> ProductList , Firebase ref){
-        info.hoang8f.widget.FButton incrButton = (info.hoang8f.widget.FButton) v.findViewById(R.id.addbutton);
-        info.hoang8f.widget.FButton qtyButton = (info.hoang8f.widget.FButton) v.findViewById(R.id.quantitybutton);
+        info.hoang8f.widget.FButton incrButton = v.findViewById(R.id.addbutton);
+        info.hoang8f.widget.FButton qtyButton = v.findViewById(R.id.quantitybutton);
+
 
         ProductClass Product = ProductList.get(position);
         // WE HAVE DEFINED NEW VARIABLE FOR CART PRODUCT AS WE ALSO HAVE TO ADD QUANTITY TO CART
 
         ProductClass CartProduct = new ProductClass(Product.getName()
                 ,Product.getImage()
-                ,Product.getprice()
+                ,Product.getPrice()
                 ,Product.getDescription()
                 ,Product.getSeller()
                 ,Product.getId()
@@ -875,7 +924,7 @@ System.out.println("The read failed: " + firebaseError.getMessage());
        if((qty+1) <= Product.getLimit()) {
             CartProduct.setQuantity(++qty);
             Product.setQuantity(qty);
-            /** PUSHING THE PRODUCT TO THE CART DATABASE **/
+           /* PUSHING THE PRODUCT TO THE CART DATABASE **/
             ref.child(CartProduct.getId()).setValue(CartProduct);
             //THIS FUNCTION WILL UPDATE THE CART COUNTER NODE IN OUR DATABASE
             updateCartCounterInFirebase();
@@ -887,11 +936,12 @@ System.out.println("The read failed: " + firebaseError.getMessage());
          //   customToastDisplay(UserActivity.this,getResources().getString(R.string.QuantityLimitExceeded));
         }
 
+        //throw new RuntimeException("Test Crash"); // Force a crash
     }
 
     public void subItemButtonClick(int position, View v, ArrayList<ProductClass> ProductList , Firebase ref){
-        TextView addedInfo = (TextView) v.findViewById(R.id.addedInfo);
-        info.hoang8f.widget.FButton qtyButton = (info.hoang8f.widget.FButton) v.findViewById(R.id.quantitybutton);
+        TextView addedInfo = v.findViewById(R.id.addedInfo);
+        info.hoang8f.widget.FButton qtyButton = v.findViewById(R.id.quantitybutton);
         //  addedInfo.setVisibility(View.VISIBLE);
         //  addedInfo.setText("Reduced One Item");
 
@@ -899,7 +949,7 @@ System.out.println("The read failed: " + firebaseError.getMessage());
         // WE HAVE DEFINED NEW VARIABLE FOR CART PRODUCT AS WE ALSO HAVE TO ADD QUANTITY TO CART
         ProductClass CartProduct = new ProductClass(Product.getName()
                 ,Product.getImage()
-                ,Product.getprice()
+                ,Product.getPrice()
                 ,Product.getDescription()
                 ,Product.getSeller()
                 ,Product.getId()
@@ -910,14 +960,14 @@ System.out.println("The read failed: " + firebaseError.getMessage());
 
         CartProduct.setQuantity(--qty);
         Product.setQuantity(qty);
-        /** PUSHING THE PRODUCT TO THE CART DATABASE **/
+        /* PUSHING THE PRODUCT TO THE CART DATABASE **/
         ref.child(CartProduct.getId()).setValue(CartProduct);
         //THIS FUNCTION WILL UPDATE THE CART COUNTER NODE IN OUR DATABASE
         updateCartCounterInFirebase();
-        info.hoang8f.widget.FButton addCart = (info.hoang8f.widget.FButton) v.findViewById(R.id.Addtocart);
-        info.hoang8f.widget.FButton addItem = (info.hoang8f.widget.FButton) v.findViewById(R.id.addbutton);
-        info.hoang8f.widget.FButton subItem = (info.hoang8f.widget.FButton) v.findViewById(R.id.subbutton);
-        info.hoang8f.widget.FButton quantity = (info.hoang8f.widget.FButton) v.findViewById(R.id.quantitybutton);
+        info.hoang8f.widget.FButton addCart = v.findViewById(R.id.Addtocart);
+        info.hoang8f.widget.FButton addItem = v.findViewById(R.id.addbutton);
+        info.hoang8f.widget.FButton subItem = v.findViewById(R.id.subbutton);
+        info.hoang8f.widget.FButton quantity = v.findViewById(R.id.quantitybutton);
 
         if( qty == 0) {
             addItem.setVisibility(View.GONE);
@@ -927,7 +977,7 @@ System.out.println("The read failed: " + firebaseError.getMessage());
         }
         else {
             quantity.setText(CartProduct.getQuantity().toString());
-            info.hoang8f.widget.FButton incrButton = (info.hoang8f.widget.FButton) v.findViewById(R.id.addbutton);
+            info.hoang8f.widget.FButton incrButton = v.findViewById(R.id.addbutton);
             incrButton.setEnabled(true);
 
         }
@@ -938,22 +988,19 @@ System.out.println("The read failed: " + firebaseError.getMessage());
         getSupportActionBar().setTitle(title);
     }
 
-    public void setFlags ( boolean pendingOrderFlag, boolean placedOrderFlag, boolean deliveredOrderFlag) {
+    /*public void setFlags ( boolean pendingOrderFlag, boolean placedOrderFlag, boolean deliveredOrderFlag) {
         this.pendingOrderFlag = pendingOrderFlag;
         this.placedOrderFlag = placedOrderFlag;
         this.deliveredOrderFlag = deliveredOrderFlag;
-    }
+    }*/
     // Sort Orderslary by Date
-    Comparator<OrderClass> orderClassComparator = new Comparator<OrderClass>() {
-        @Override
-        public int compare(OrderClass e1, OrderClass e2) {
-            if(e1.getDate().getTime() > e2.getDate().getTime()) {
-                return -1;
-            } else if (e1.getDate().getTime() < e2.getDate().getTime()) {
-                return 1;
-            } else {
-                return 0;
-            }
+    Comparator<OrderClass> orderClassComparator = (e1, e2) -> {
+        if(e1.getDate().getTime() > e2.getDate().getTime()) {
+            return -1;
+        } else if (e1.getDate().getTime() < e2.getDate().getTime()) {
+            return 1;
+        } else {
+            return 0;
         }
     };
 }
